remove output to own package
